requirements.txt
Jinja2==2.6
SQLAlchemy==0.6.4
Werkzeug==0.8.1
lxml==2.3.1
ming
nose
pytest
forked-path
-e git+git://github.com/Singletoned/validino.git#egg=validino
-e git+git://github.com/Singletoned/testino.git#egg=testino
-e .
